Refactor App and Home components to use OpenFeature React SDK for feature flag management
